clean and order the code
