+++ TableClient/src/App.css
@@ -1,5 +1,4 @@
 #root {
-  /* max-width: 1280px; */
   text-align: center;
 }
