+++ TableClient/src/components/filters/GeneralFilter.tsx
