+++ TableClient/src/components/users/UpdateUserDetails.tsx
@@ -1,15 +1,19 @@
-import { useContext, useEffect, useState } from "react";
+import React, { useContext, useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { ServerContext } from "../context/ServerUrlContext";
-import { resetPassword } from "../api/userApi";
+import { ServerContext } from "../../context/ServerUrlContext";
+import { UserContext } from "../../context/userContext";
+import { updateUserDetails } from "../../api/userApi";
 
-function ResetPassword() {
+export default function UpdateUserDetails() {
+  const {user} = useContext(UserContext) //get the userName from userContext
+  const [newUserName, setNewUserName] = useState<string>(user);
   const [password, setPassword] = useState<string>("");
   const [confirmPassword, setConfirmPassword] = useState<string>("");
   const [visible, setVisible] = useState(false);
   const [visibleConfirm, setVisibleConfirm] = useState(false);
   const [match, setMatch] = useState(false);
-  const [email, setEmail] = useState<string>("");
+  const {email} = useContext(UserContext)
+  const [newEmail, setNewEmail] = useState<string>(email);
   const [timeoutId, setTimeoutId] = useState<number | null>(null);
   const serverUrl = useContext(ServerContext);
   const navigate = useNavigate();
@@ -52,20 +56,19 @@ function ResetPassword() {
     validate();
   }, [confirmPassword, password]);
 
-  const handleSubmitResetPassword = async (
+  const handleSubmitUpdateUserDetails = async (
     ev: React.FormEvent<HTMLFormElement>
   ) => {
     try {
       ev.preventDefault();
-      console.log("At handleSubmitRegister, the serverUrl is:", serverUrl);
+      console.log("At handleSubmitUpdateUserDetails, the serverUrl is:", serverUrl);
       if (password === confirmPassword) {
-        const data = { email, password };
-        if (!data) throw new Error("register failed - no email or password");
-        const response = await resetPassword({ serverUrl, email, password });
-        // const response = await register({userName, password});
+        const data = { newUserName, newEmail, password };
+        if (!data) throw new Error("update detail failed - no userName, email or password");
+        const response = await updateUserDetails({ serverUrl, email: newEmail, password });
         if (!response) throw new Error("register failed from server");
-        alert("Password successfully set");
-        navigate("/");
+        alert("User details updated successfully");
+        navigate("/userPage");
       } else {
         alert("Passwords not match, please check it out");
       }
@@ -79,25 +82,42 @@ function ResetPassword() {
       <div className="h-screen">
         <button
           className="absolute top-8 left-16"
-          onClick={() => navigate("/")}
+          onClick={() => navigate("/userPage")}
         >
           Back
         </button>
 
         <form
           className="custom-form relative"
-          onSubmit={handleSubmitResetPassword}
+          onSubmit={handleSubmitUpdateUserDetails}
         >
-        <h1 className="text-7xl mb-20 font-bold mt-16 text-black">Set a New Password</h1>
+          <h1 className="text-7xl mb-20 font-bold mt-16 text-black">
+            Update Your Details
+          </h1>
+
+          <div className="">
+            <div className="icon_container"></div>
+            <input
+              className="border border-black m-2 rounded-2xl w-72 h-12 relative indent-4"
+              type="userName"
+              name="userName"
+              autoComplete="given-name"
+              placeholder={user}
+              value={newUserName}
+              onInput={(ev) =>
+                setNewUserName((ev.target as HTMLInputElement).value)
+              }
+            />
+          </div>
 
           <input
             className="border border-black m-2 rounded-2xl w-72 h-12 relative indent-4"
             type="email"
             name="email"
             autoComplete="given-name"
-            placeholder="Enter your email" //must be uniq
-            value={email}
-            onInput={(ev) => setEmail((ev.target as HTMLInputElement).value)}
+            placeholder={email}
+            value={newEmail}
+            onInput={(ev) => setNewEmail((ev.target as HTMLInputElement).value)}
           />
 
           <div className="relative left-4">
@@ -107,12 +127,13 @@ function ResetPassword() {
               id="password"
               name="password"
               autoComplete="off"
-              placeholder="Choose your Password"
+              placeholder="Password"
               value={password}
               onInput={(ev) =>
                 setPassword((ev.target as HTMLInputElement).value)
               }
             ></input>
+  
             <button
               type="button"
               className="emoji-button visible right-12"
@@ -131,11 +152,12 @@ function ResetPassword() {
               value={confirmPassword}
               name="confirmPassword"
               autoComplete="off"
-              placeholder="Enter your Password again"
+              placeholder="Enter Password again"
               onInput={(ev) =>
                 setConfirmPassword((ev.target as HTMLInputElement).value)
               }
             ></input>
+
             <button
               type="button"
               className="emoji-button visible right-12"
@@ -148,12 +170,10 @@ function ResetPassword() {
 
           <p>{!match ? "password are not matched!" : null}</p>
           <button className="registerBtn text-xl" type="submit">
-            Save new password
+            Save
           </button>
         </form>
       </div>
     </div>
   );
 }
-
-export default ResetPassword;
