+++ TableClient/src/index.css
@@ -19,7 +19,6 @@ a:hover {
 }
 
 body {
-  /* display: flex; */
   place-items: center;
   min-width: 320px;
   min-height: 100vh;
@@ -50,10 +49,6 @@ button:focus-visible {
 }
 
 @media (prefers-color-scheme: light) {
-  /* :root {
-    color: #213547;
-    background-color: #ffffff;
-  } */
   a:hover {
     color: #747bff;
   }
