+++ TableClient/src/router/router.tsx
@@ -1,8 +1,8 @@
 import { createBrowserRouter } from "react-router-dom"
-import ForgotPassword from "../components/ForgotPassword"
-import OTPInput from "../components/OTPInput"
-import ResetPassword from "../components/ResetPassword"
-import UpdateUserDetails from "../components/UpdateUserDetails"
+import ForgotPassword from "../components/users/password/ForgotPassword"
+import OTPInput from "../components/users/password/OTPInput"
+import ResetPassword from "../components/users/password/ResetPassword"
+import UpdateUserDetails from "../components/users/UpdateUserDetails"
 import LandingPage from "../view/LandingPage"
 import MainTablesPage from "../view/MainTablesPage"
 import RegisterPage from "../view/RegisterPage"
