--- TableClient/src/style/mainTablePage.css
@@ -1,3 +1,0 @@
-/* body {
-    display: block;
-  } */
