--- TableClient/src/style/oldTableData copy.css
@@ -1,34 +1,0 @@
-/* Style for the table headers */
-th {
-    position: relative;
-    padding: 10px;
-    background-color: #f8f8f8;
-    border: 1px solid #ddd;
-    text-align: left;
-  }
-  
-  /* Style for the "+" button */
-  .plus-button {
-    position: absolute;
-    top: 50%;
-    right: -12px; /* Adjust position to place it between columns */
-    transform: translateY(-50%);
-    background-color: #f0f0f0;
-    border: 1px solid #ccc;
-    border-radius: 50%;
-    width: 24px;
-    height: 24px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    cursor: pointer;
-    opacity: 0;
-    transition: opacity 0.2s;
-    z-index: 1;
-  }
-  
-  /* Show the "+" button when hovering over the corresponding header cell */
-  th:hover .plus-button {
-    opacity: 1;
-  }
-
