--- TableClient/src/style/toast.css
@@ -1,13 +1,0 @@
-.toast-container {
-    width: 100%;
-    max-width: 300px;
-}
-
-.toast-message-loading {
-    background-color: #333;
-    color: white;
-    padding: 16px;
-    border-radius: 8px;
-    text-align: center;
-    font-size: 16px;
-}
